+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Send username to LTI Consumer if user is logged in
 - Make related courses optional for program page
 
 ## [2.6.0] - 2021-05-03